pyproject.toml
[tool.poetry]
name = "ontology-loader"
version = "0.1.0"
description = "Import an ontology file into a Neo4j database"
authors = ["fcriscuo <fcriscuo@genomicsai.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rdflib = "^7.0.0"
regex = "^2023.10.3"
pyparsing = "^3.1.1"
pronto = "^2.5.5"
neo4j = "^5.15.0"
google-generativeai = "^0.3.2"
sentence-transformers = "^2.4.0"
huggingface-hub = "^0.21.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
